pyproject.toml
[project]
name = "vocabmaster"
version = "2.0.0"
description = "词汇测试系统 v2.0 - 具备增强缓存、性能监控、完整IELTS词汇库等功能，适用于 macOS 本地开发，使用 Poetry 管理依赖。"
authors = [
    {name = "Icarus603",email = "zhehongl91@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pyqt6 (>=6.6.0,<6.8.0)",
    "numpy (>=1.24.0,<2.0.0)", 
    "requests (>=2.31.0,<3.0.0)",
    "scikit-learn (>=1.3.0,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pyinstaller = "5.13.2"
